# swagger.yaml
openapi: 3.0.0
info:
  title: Keijiban API
  description: Incredible Keijiban!
  version: 0.0.1
servers:
  - url: http://localhost
    description: For development.
paths:
...
  /api/user/profile:
    put:
      summary: Update a user profile (username and email).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                username:
                  type: string
                email:
                  type: string
                email_old:
                  type: string
      responses:
        "200":
          description: A user object.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
                  email:
                    type: string
                  crated_at:
                    type: string
        default:
          description: Unexpected error.
  /api/user/password:
    put:
      summary: Update a user password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                password:
                  type: string
      responses:
        "200":
          description: A user object.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
                  email:
                    type: string
                  crated_at:
                    type: string
        default:
          description: Unexpected error.
  /api/user/{id}:
    get:
      summary: Return a user by ID.
      parameters:
        - name: id
          in: path
          required: true
          description: user ID
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Return a user object.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
                  email:
                    type: string
                  created_at:
                    type: string
        default:
          description: Unexpected error.
    delete:
      summary: Deletes a user by ID.
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the user to delete.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: A user object.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
                  email:
                    type: string
                  crated_at:
                    type: string
        default:
          description: Unexpected error.
...